chore(coordinator): apply review cleanups
Consolidate duplicate proof index providers, remove confirmed dead APIs, and make atomic holder references immutable.

Refs #3635

Refs #3686
