+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/l1/Web3JLineaValidiumSmartContractClient.kt
@@ -125,26 +125,6 @@ class Web3JLineaValidiumSmartContractClient(
       }
   }
 
-  override fun finalizeBlocksEthCall(
-    aggregation: ProofToFinalize,
-    aggregationLastBlob: BlobRecord,
-    parentL1RollingHash: ByteArray,
-    parentL1RollingHashMessageNumber: Long,
-  ): SafeFuture<String?> {
-    return getVersion()
-      .thenCompose { version ->
-        val function =
-          Web3JLineaValidiumFunctionBuilders.buildFinalizeBlocksFunction(
-            version,
-            aggregation,
-            aggregationLastBlob,
-            parentL1RollingHash,
-            parentL1RollingHashMessageNumber,
-          )
-        web3jContractHelper.executeEthCall(function)
-      }
-  }
-
   override fun finalizeBlocksAfterEthCall(
     aggregation: ProofToFinalize,
     aggregationLastBlob: BlobRecord,
