+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/l1/Web3JLinethRollupSmartContractClient.kt
@@ -171,26 +171,6 @@ class Web3JLinethRollupSmartContractClient internal constructor(
       }
   }
 
-  override fun finalizeBlocksEthCall(
-    aggregation: ProofToFinalize,
-    aggregationLastBlob: BlobRecord,
-    parentL1RollingHash: ByteArray,
-    parentL1RollingHashMessageNumber: Long,
-  ): SafeFuture<String?> {
-    return getVersion()
-      .thenCompose { version ->
-        val function =
-          Web3JLinethRollupFunctionBuilders.buildFinalizeBlocksFunction(
-            version,
-            aggregation,
-            aggregationLastBlob,
-            parentL1RollingHash,
-            parentL1RollingHashMessageNumber,
-          )
-        web3jContractHelper.executeEthCall(function)
-      }
-  }
-
   override fun finalizeBlocksAfterEthCall(
     aggregation: ProofToFinalize,
     aggregationLastBlob: BlobRecord,
